+++ project-editor/src/renderer/components/EditorComponents/ObjectListItem.vue
@@ -1,5 +1,5 @@
 <template>
-    <li class="obj_list_item" :class="{ selected: this.isObjectSelected }" @click="select">
+    <li class="obj_list_item" :class="{ selected: isObjectSelected() }" @click="select">
         <div class="icon">
             <img src="static/images/rectangleIcon.png" alt="Rectangle icon" v-if="type === 'rectangle'">
             <img src="static/images/textIcon.png" alt="Text icon" v-else-if="type === 'text'">
@@ -17,20 +17,18 @@ import { mapGetters, mapActions } from 'vuex';
 import '../../store/index.js';
 
 export default {
-    computed: {
-        ...mapGetters(['selectedObjects']),
-
-        isObjectSelected: function() {
-            return this.selectedObjects.indexOf(this.id) !== -1;
-        }
-    },
+    computed: mapGetters(['selectedObjects']),
     methods: {
         ...mapActions(['selectObject', 'unselectObject', 'clearSelectedObjects']),
         
-        select: function(event) {
+        isObjectSelected() {
+            return this.selectedObjects.indexOf(this.id) !== -1;
+        },
+
+        select(event) {
             const isCtrlKey = event.ctrlKey;
 
-            if(!this.isObjectSelected) {
+            if(!this.isObjectSelected()) {
                 for(let i in this.selectedObjects) {
                     this.unselectObject({ 
                         from: i, 
@@ -49,9 +47,6 @@ export default {
                 }
             }
         }
-    },
-    created: function() {
-        this.clearSelectedObjects();
     },
     props: {
         name: String,